Fix no_std build errors caused by SerializeSeed impls
